# Conflicts:
#	Next2/src/Next2/Constants.cs
#	Next2/src/Next2/Next2.csproj
#	Next2/src/Next2/Resources/Strings/Strings.resx

+++ Next2/src/Next2/Models/CustomerBindableModel.cs
@@ -21,7 +21,7 @@ namespace Next2.Models
 
         public int Points { get; set; }
 
-        public double GiftCardTotal { get; set; }
+        public float GiftCardTotal { get; set; }
 
         public ImageSource CheckboxImage { get; set; }
